Ignore anchors/queryparams when navigating
A [recent helpdesk ticket][1] alerted us to a bug in bramble where navigation to certain urls fails - in particular, urls with anchors or query params.

```
<!-- These work properly -->
<a href="page2.html">To page 2</a>
<a href="#end-of-page">To end of page 1</a>

<!-- These are broken -->
<a href="page2.html#end-of-page">End of page 2</a>
<a href="page2.html?q=a">Page 2 with queryparams</a>
```

In the broken cases the links are simply unresponsive - a message is generated in the browser console and no navigation occurs.  Note that these work fine in shared projects when we load the source normally and let the browser handle navigation.

Fixing these properly in the bramble editing environment (which has manual handlers for both navigation and anchors) is fairly involved, and we've scheduled it further out.  This commit is a partial mitigation that allows navigation to succeed for the currently failing cases, by completely ignoring the anchor and/or queryparam portions of the URL.

[1]: https://codeorg.zendesk.com/agent/tickets/110284